Make major changes for v2
In v2 I've added a new function "whatis". Everything about it will be 
mentioned in README file. Next up I've changed sort to filter and 
removed it from index.js since it is irrelevant to the package. Also 
added comments to spice up the code & on top of that 
ERR_INVALID_ARG_TYPE & bindChecker have callbacks (few improvements as 
well).